디렉토리 main으로 병합, 프롬프트 세부 설정 추가
