--- prompts/prompt.py
@@ -1,183 +1,0 @@
-def get_summary_prompt(type_name, content):
-    prompts = {
-        "내용 요약_기본 요약": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 요약 전문가로, 복잡한 개념도 학습자의 눈높이와 학습 목적에 맞는 콘텐츠로 재구성하는 데 특화되어 있다. ",
-            "user": f'''문서 전체를 분석하여 문서의 흐름을 파악하고, 아래의 예시와 같이 '서론-본론'결론'으로 구조화하며, 정리한 각 구간은 아래의 예시와 같이 '구간 제목-구간 핵심 요약(2-3문장)'의 형식으로 요약하라.\n\n
-                        #예시\n
-                        [서론 - 탄소중립의 정의와 등장 배경]\n
-                        - 탄소중립은 배출한 온실가스를 다시 흡수하거나 제거함으로써 실질적 배출량을 0으로 만드는 개념으로, 기후 변화 대응을 위한 국제적 목표로 부각되고 있다.\n  
-                        [본론 - 실천 분야와 구체적 전략]\n  
-                        - 탄소중립은 에너지 전환, 친환경 교통 시스템, 건물의 에너지 효율화 같은 다양한 기술 및 정책 변화로 실현되고 있다.\n  
-                        [결론 - 탄소중립의 확장적 의미와 필요성]\n  
-                        - 탄소중립은 단순한 환경 보호를 넘어 경제 경쟁력 강화, 사회 정의 실현 등 구조적 전환을 수반하는 종합적 과제로 인식되고 있다.\n\n  
-                        문서에서 가장 중요한 5-10개의 핵심 개념 또는 키워드를 선별하고, 각 개념이 어떤 위치에 등장하며, 어떤 정보를 연결하거나 확장하며, 전체 문장의 흐름 속에서 어떤 기능을 수행하는지에 대해 아래와 같이 '키워드-내용'의 형식으로 요약하라.\n\n
-                        #예시\n
-                        탄소중립 - 문서 전체를 관통하는 중심 개념으로, 기후 변화에 대응하기 위한 국제적 전략이자 사회 전반의 전환을 이끄는 방향성을 가진다.\n\n  
-                        이제까지 분석한 문서의 구조와 핵심 개념을 활용해, [고등학생] 수준의 학습자가 쉽게 이해할 수 있도록 문서 전체 내용을 분석하여 핵심 내용을  '서론-본론-결론'의 형식으로 [500]자 이내로 요약문을 재구성하라.
-                        \n\n\n{content}'''
-        },
-        "내용 요약_핵심 요약": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 전문가로, 다양한 분야의 방대한 문서에서 핵심만 추출하여 정보를 압축하는 데 특화되어 있다.",
-            "user": f'''이 문서는 전체 내용 중 가장 중요한 핵심 아이디어를 하나 선별하여 본질적인 핵심 내용만을 정제하여 전달함으로써 [고등학생] 수준의 학습자가 문서의 핵심을 빠르게 이해할 수 있도록 한다.\n\n
-                        ‘서론-본론-결론’ 등의 구분 없이 문서의 본질만을 아래의 예시와 같이 단일 문장으로 요약하라.\n
-                        ##예시\n
-                        탄소중립은 기후 위기 대응을 위한 국제적 전략으로, 온실가스 감축을 넘어 산업·사회 전반의 구조 전환을 요구하는 핵심 개념이다.\n\n
-                        문서의 주제나 목적을 가장 잘 상징하는 핵심 개념을 1개 추출하고, 그 개념이 문서 내에서 어떤 의미,역할, 문맥적 특징으로 동작하는지 대해 아래의 예시와 같이 1-2문장 이내로 간결하게 요약하라.\n
-                        ##예시\n
-                        탄소중립 - 문서의 중심 개념으로, 환경 문제 해결을 넘어서 에너지 구조, 산업 경쟁력, 사회 정의를 포괄하는 통합적 전환 프레임으로 기능하며, 정책과 기술 변화의 방향을 제시하는 핵심 축이다.\n\n
-                        이제까지 분석한 문서의 구조와 핵심 개념을 바탕으로, 문서의 핵심 내용을 200자 이내의 1-2문장으로 요약하라.\n
-                        \n\n{content}'''
-        },
-        "내용 요약_주제 요약": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 전문가로, 다양한 문서에서 중심 주제를 빠르게 파악하고 요약하는 데 특화된 교육 콘텐츠 기획자다.",
-            "user": f'''이 문서는 복잡하게 얽힌 정보를 주요 주제 단위로 나누어 정리함으로써 [요약수준] 수준의 학습자가 핵심 내용을 구조적으로 이해할 수 있도록 해야 한다.\n
-                        문서 전체를 분석하여 중심이 되는 주제들을 식별하고, 각 주제별로 핵심 내용을 정리하여 정리한 각 주제에 대해, 문서 내에서 그 주제가 어떤 의미를 가지며, 왜 핵심적인지를 1-2문장 이내로 아래의 예시와 같이 '주제 제목 - 요약' 형식으로 요약하라.\n\n
-                        #예시\n
-                        [주제1] 탄소중립의 개념과 글로벌 추진 배경\n
-                        - 이 주제는 문서 전반이 출발점이자 전체 내용을 이끄는 핵심 개념으로, 탄소중립의 정의와 국제적 필요성을 설명한다. \n\n
-                        분석된 문서의 구조를 바탕으로 중심 내용을 대표할 수 있는 핵심 주제를 [5]개를 도출하라.\n
-                        불필요하게 일반적인 단어나 표면적 문장은 피하고, 내용의 본질과 중심성에 기반한 판단을 수행하라.\n\n
-                        이제까지 분석된 핵심 주제를 바탕으로, 선정된 핵심 주제 각각에 대해 요약하라.\n
-                        각 요약은 [분야]를 기반으로 하고, 단순 정보 압축이 아닌 [고등학생] 수준의 학습자가 이해할 수 있도록 문서의 전체 흐름 속에서 해당 주제가 어떤 의미를 가지는지를 [500]자 이내로 작성하라.\n
-                        \n\n{content}'''
-        },
-        "내용 요약_목차 요약": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 요약 전문가로, 복잡한 문서를 단원 또는 목차 구조에 따라 분석하고 요약하는 데 특화되어 있다.",
-            "user": f'''문서의 구성과 흐름을 정확히 파악하고, [분야] 수준의 학습자가 쉽게 이해할 수 있도록 핵심 내용을 구조적으로 재구성하라.\n\n
-                        이 문서는 전체 내용을 구성 단위별로 나누고, 각 단위의 핵심 내용을 요약함으로써 [고등학생] 수준의 학습자가 문서 전체의 흐름과 주요 내용을 쉽게 이해할 수 있어야 한다.\n\n
-                        문서를 전체를 분석하여 목차별(또는 논리적 구간별) 구조를 다음 조건에 따라 구조화하라.\n
-                        문서에 명시된 목차가 있는 경우, 그 목차 체계를 그대로 따라 각 구간(장, 절, 항 등)을 기준으로 나누고, 문서에 명시된 목차가 없는 경우, 제목/소제목/문단의 논리 흐름을 바탕으로 내용을 '도입-배경-전개-결론'의 구간으로 나누어라.\n
-                        정리한 각 구간은 아래의 예시와 같이 '구간 제목-구간 핵심 요약(2-3문장)'의 형식으로 요약하라.\n\n
-                        ##예시\n
-                        [도입 - 탄소중립의 개념 정의]\n
-                        - 탄소중립은 온실가스 배출량과 흡수량의 균형을 통해 실질적인 순배출량을 0으로 만드는 개념으로, 기후 변화 대응의 핵심 전략으로 등장했다. \n
-                        [배경 - 글로벌 추진과 참여 주체]\n
-                        - 전 세계적으로 탄소중립 목표가 설정되면서, 정부, 기업, 개인이 모두 역할을 수행해야 하는 새로운 사회적 요구가 형성되었다.\n  
-                        [전개 - 분야별 실천 전략]\n  
-                        - 에너지 전환, 친환경 교통, 건물 에너지 효율화 등 다양한 부문에서 기술적·정책적 변화가 동시에 추진되고 있다.\n  
-                        [결론 - 구조적 전환의 필요성과 전망]\n  
-                        - 탄소중립은 단순한 환경 과제가 아니라 경제 구조, 산업 경쟁력, 사회 정의까지 포함하는 전방위적 전환을 요구한다.\n\n  
-                        분석된 문서의 구조를 바탕으로, 해당 구간의 중심 개념 또는 키워드를 1-2개씩 추출하고, 각 키워드에 대해 왜 중요한지 또는 어떤 기능을 하는지를 아래의 예시와 같이 '키워드 -설명'의 형식으로 정리하라.\n
-                        ##예시\n
-                        탄소중립 - 온실가스의 순배출을 0으로 만들어 기후 위기를 억제하려는 핵심 정책 개념이자, 모든 전개 내용의 출발점이다.\n
-                        분석된 문서 구조와 각 구간별 핵심 개념을 바탕으로, 긱 구간별 핵심 내용을 요약하라.\n
-                        각 요약은 [분야]를 기반으로 하고,  독립적으로 이해 가능해야 하며, 단순 정보 압축이 아닌 [고등학생] 수준의 학습자가 이해할 수 있도록 문서의 핵심 흐름과 메시지가 유기적으로 연결하여 [500]자 이내로 작성하라.\n
-                        \n\n{content}'''
-        },
-        "내용 요약_키워드 요약": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 키워드 중심 콘텐츠 요약 전문가다. 복잡한 문서 속에서 지정된 키워드와 직접적으로 연결된 정보만을 추출하고, 학습 목적에 맞게 간결하고 논리적으로 재구성하는 데 특화되어 있다.",
-            "user": f'''이 문서는 학습자가 지정한 키워드가 등장하는 문맥과 연결된 내용을 파악하고, 관련 정보만을 선별하여 요약함으로써 [요약수준] 수준의 학습자가 키워드 기반 문서의 핵심 내용을 쉽게 이해할 수 있어야 한다.\n\n
-                        학습자가 지정한 키워드(최대 3개, 각 10자 미만)을 기준으로 문서 내 흐름과 구조 속에서 키워드의 위치와 정보 구획을 파악하고, 아래의 예시와 같이 '키워드-요약'형식으로 요약하라.\n\n
-                        ##예시\n
-                        [에너지 - 요약]\n
-                        - 에너지 전환은 탄소중립 실현의 핵심 요소로, 화석연료 중심에서 재생에너지 중심으로의 구조 변화가 요구된다.\n\n 
-                        분석된 문서의 구조를 바탕으로, 각 키워드가 문서 내에서 어떤 의미, 역할, 문맥적 특징, 중심 개념으로 작동하는지에 대해 아래의 예시와 같이 1-2문장 이내로 간결하게 요약하라.\n
-                        ##예시\n
-                        에너지 - 탄소중립 실현을 위한 구체적 실천 수단이자, 기술적 전환의 핵심 경로로 작용한다.\n\n  
-                        이제까지 분석된 문서의 구조와 선별한 핵심 개념에 대한 핵심 내용을 요약하라.\n
-                        각 요약은 [분야]를 기반으로 하고, [고등학생] 수준의 학습자가 이해할 수 있도록 핵심 메시지를 중심으로 간결하고 논리적으로 구성하여 [500]자 이내로 요약하라.\n
-                        \n\n{content}'''
-        },
-        "문제 생성_n지 선다형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 객관식 문제를 설계하는 데 특화되어 있다.",
-            "user": f'''학습자의 핵심 개념 이해, 용어 정의, 개념 간 비교 및 오류 식별 능력을 점검할 수 있도록 [분야]와 관련한 [4]지선다형 문제를 생성하라.\n
-                        혼동 가능한 오답을 포함해 학습자의 개념적 정확성과 구분 능력을 평가할 수 있도록 구성하라.\n
-                        문제 출력 형식은 아래의 예시와 같이 구성하라.\n\n
-                        ##문제 생성 예시\n
-                        [문제]\n
-                        [보기]\n 
-                        [정답]\n
-                        [해설]\n\n
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        보기는 반드시 [4]개만 생성하라.\n
-                        문제의 문장 구조는 한 형식으로만 고정하지 말고 다양한 구조를 활용하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_순서 배열형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 제시된 선택지를 올바른 순서대로 배열함으로써 개념의 구조와 전개 흐름을 논리적으로 이해했는지를 점검할 수 있도록 [분야]와 관련한 배열 문제를 구성하라.\n
-                        문제 출력 형식은 아래의 예시와 같이 구성하라.\n
-                        ##문제 생성 예시\n\n
-                        [문제]\n
-                        [보기]\n
-                        [정답]\n
-                        [해설]\n
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        보기는 반드시 [4]개만 생성하라.\n
-                        문제의 문장 구조는 "다음 보기들을 순서대로 배열하시오." 등 예시의 형식으로만 고정하지 말고, 반드시 다양한 구조를 활용하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_참거짓형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 문장에 포함된 정보가 정확한지, 오개념을 구별할 수 있는지 등을 검토할 수 있도록 [분야]와 관련한 참거짓형 문제를 생성하라.\n
-                        문제 출력 형식은 아래의 예시와 같이 구성하라.\n\n 
-                        ##문제 생성 예시\n 
-                        [문제]\n 
-                        [보기]\n 
-                        [정답]\n 
-                        [해설]\n\n 
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n 
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        [문제]은 반드시 "다음 진술문을 읽고, 옳으면 O, 틀리면 X로 표시하시오."로 출력하라.\n 
-                        이외 다른 항목과 절대 결합하지 마라.\n 
-                        [보기]는 "진술문 (O/X)"의 형태로 구성하고, 반드시 하나만을 출력하라. 그 이상을 절대 출력하지 마라.\n 
-                        [정답]은 반드시 O 또는 X로만 표기하며, 설명이나 다른 문장과 절대 결합하지 마라.
-                        \n\n{content}'''
-        },
-        "문제 생성_빈칸 채우기형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 문맥 속 개념이나 핵심 용어를 정확히 이해하고 적용할 수 있는지를 점검할 수 있도록 [분야]와 관련한 빈칸 채우기형 문제를 생성하라.\n
-                        문제 출력 형식은 아래의 예시와 같이 구성하라.\n\n
-                        [지시문]\n
-                        [보기]\n
-                        [정답]\n
-                        [해설]\n
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        각 문제는 빈칸을 반드시 [1]개만 생성하라.\n
-                        [지시문]은 반드시 "다음 문장을 읽고, 빈칸에 알맞은 용어를 쓰시오"만 출력하라. 이외 다른 항목과 절대 결합하지 마라.\n
-                        [보기]는 문제 당 하나의 문단 또는 문장으로 구성된 지문이며, 문맥상 핵심 개념이 빠진 빈칸을 포함한다.\n
-                        빈칸은 (1)__, (2)__의 형식으로 작성하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_단답형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''문제 출력 형식은 아래의 예시와 같이 구성하라.\n\n
-                        [문제]\n   
-                        [정답]\n
-                        [해설]\n
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        [문제]의 문장 구조는 "~은 무엇인가?", "~을 쓰시오.", "~에 해당하는 개념은?" 등 예시의 형식으로만 고정하지 말고, 반드시 다양한 구조를 활용하라.\n
-                        [정답]은 문장이 아닌 단어나 짧은 구나 숫자로 구성하고, 용어, 개념, 키워드 중심으로 마침표 없이 간결하게 작성한다.
-                        \n\n{content}'''
-        },
-        "문제 생성_서술형": {
-            "system": "너는 [분야]에서 20년 경력을 지닌 평가 설계 전문가로, [전공자] 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 핵심 개념, 절차, 원리를 자신의 언어로 정확하고 일관되게 설명하는 능력을 평가할 수 있도록 [분야]와 관련한 서술형 문제를 구성하라.\n
-                        문제 출력 형식은 아래의 예시와 같이 구성하라.\n
-                        [문제]\n
-                        [정답]\n
-                        [해설]\n\n
-                        아래는 문제 생성 시 따라야 할 작성 기준이다.\n
-                        문제의 수는 반드시 [3]개만 생성하라.\n
-                        문제의 난이도는 반드시 [고등학교]을 기준으로 출제하라.\n
-                        [문제]의 문장 구조는 "~의 개념과 주요 특징을 서술하시오.?", "~에 대한 자신의 견해를 서술하시오.", "~의 차이점을 서술하시오?" 등 예시의 형식으로만 고정하지 말고, 반드시 다양한 구조를 활용하라.\n
-                        [정답]은 학습자가 자신의 응답과 비교하여 핵심 개념 이해, 논리적 설명력, 구조화 수준을 점검할 수 있도록 구성된 피드백형 예시로 제시하라.\n
-                        - 문제의 요구에 따라 개념의 정의, 특징, ,절차, 비교, 사례 등을 명확하게 서술하라.\n
-                        - 학습자의 수준에 맞는 표현을 사용하되, 정확성과 일관성을 유지하라.\n
-                        [해설]은 학습자의 이해를 돕기 위한 설명과 함께, 해당 문제에서 평가되는 핵심 채점 기준을 명확히 제시하라.\n
-                        - 채점 기준은 학습자가 자기 답안을 비교하며 점검할 수 있도록 항목별로 간단히 정리하라.\n
-                        - 채점 기준은 ‘반드시 포함해야 할 요소’ 또는 ‘주요 판단 기준’의 형태로 제시하라.\n
-                        - 해설은 정답의 근거, 관련 개념 간의 연결, 오개념과의 구분 등을 포함해 학습 효과를 높이도록 구성하라.\n
-                        \n\n{content}'''
-        }
-    }
-    return prompts.get(type_name, {"system": "", "user": ""})
