--- summary/extractor.py
@@ -1,18 +1,0 @@
-import fitz  # PyMuPDF
-from pptx import Presentation
-
-def extract_text_from_pdf(file_path):
-    text = ""
-    with fitz.open(file_path) as doc:
-        for page in doc:
-            text += page.get_text()
-    return text
-
-def extract_text_from_pptx(file_path):
-    text = ""
-    prs = Presentation(file_path)
-    for slide in prs.slides:
-        for shape in slide.shapes:
-            if hasattr(shape, "text"):
-                text += shape.text + "\n"
-    return text
